results/final_count/run-20230926-130438_LLMatic-zero_shot_donnemartin-system-design-oop-whatsapp-donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_10-runs_log/dropbox/5/LLMatic/test_main.py
import pytest
from flask import url_for
from main import app

@pytest.fixture
def client():
	app.config['TESTING'] = True
	with app.test_client() as client:
		with app.app_context():
			yield client

def test_home(client):
	response = client.get('/')
	assert response.status_code == 200
	assert response.data == b'Home'

def test_register(client):
	response = client.get('/register')
	assert response.status_code == 200
	assert response.data == b'Register'

def test_login(client):
	response = client.get('/login')
	assert response.status_code == 200
	assert response.data == b'Login'

def test_upload(client):
	response = client.get('/upload')
	assert response.status_code == 200
	assert response.data == b'Upload'

def test_download(client):
	response = client.get('/download/test')
	assert response.status_code == 200

def test_share(client):
	response = client.get('/share')
	assert response.status_code == 200
	assert response.data == b'Share'

def test_encrypt(client):
	response = client.get('/encrypt')
	assert response.status_code == 200
	assert response.data == b'Encrypt'

def test_log(client):
	response = client.get('/log')
	assert response.status_code == 200

